Project/requirements.txt
boto3==1.10.34
botocore==1.13.34
Django==2.2.7
django-crispy-forms==1.8.1
django-storages==1.8
docutils==0.15.2
gunicorn==20.0.4
imutils==0.5.3
jmespath==0.9.4
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
python-dateutil==2.8.0
pytz==2019.3
s3transfer==0.2.1
six==1.13.0
sqlparse==0.3.0
gunicorn==20.0.4
urllib3==1.25.7

